save changes style
